modified files

+++ server/controllers/Section.js
@@ -1,5 +1,6 @@
 const Section = require('../models/Section');
 const Course = require('../models/Course');
+const SubSection = require("../models/SubSection");
 
 // CREATE a new section
 exports.createSection = async (req, res) => {
@@ -45,7 +46,7 @@ exports.createSection = async (req, res) => {
     }catch(error){
         return res.status(500).json({
             success: false,
-            message: 'Unable to create Section, please try again.',
+            message: 'Internal server error',
             error: error.message
         });
     }
@@ -62,9 +63,20 @@ exports.updateSection = async (req, res) => {
 			{ sectionName },
 			{ new: true }
 		);
+
+        const course = await Course.findById(courseId)
+		.populate({
+			path:"courseContent",
+			populate:{
+				path:"subSection",
+			},
+		})
+		.exec();
+
         return res.status(200).json({
             success: true,
             message: section,
+            data:course,
         });
     } catch (error) {
         console.error("Error updating section:", error);
@@ -80,19 +92,38 @@ exports.updateSection = async (req, res) => {
 exports.deleteSection = async (req, res) => {
     try{
         const { sectionId } = req.params;
+        await Course.findByIdAndUpdate(courseId, {
+			$pull: {
+				courseContent: sectionId,
+			}
+		})
+        const section = await Section.findById(sectionId);
+        console.log(sectionId, courseId);
+		if(!section) {
+			return res.status(404).json({
+				success:false,
+				message:"Section not Found",
+			})
+		}
+
+        //delete sub section
+		await SubSection.deleteMany({_id: {$in: section.subSection}});
+
         await Section.findByIdAndDelete(sectionId);
 
-        // remove section for course
-        // await Course.updateMany(
-        //     { _id: sectionId },
-        //     { $pull: { _id: sectionId }},
-        //     { new: true } 
-        // )
-        // await Course.findById({_id: courseId})
-
+        //find the updated course and return 
+		const course = await Course.findById(courseId).populate({
+			path:"courseContent",
+			populate: {
+				path: "subSection"
+			}
+		})
+		.exec();
+        
         return res.status(200).json({
             success: true,
-            message: 'Section deleted'
+            message: 'Section deleted',
+            data:course
         });
     } catch (error) {
         console.error("Error deleting section:", error);
